+++ CHANGELOG.md
@@ -6,8 +6,23 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `RocketIcon` and `RocketCircleIcon` component.
 - Feature: Create `ProgressRing` component.
 
-- Fix: Fix arrow size on `ProjectCarousel` buttons.
-- Feature: Add a11y options for `ProjectCarousel` buttons.
+## [2.64.0] - 2020-04-07
+
+Feature:
+- Add a11y options for new `Carousel` buttons.
+
+Fix:
+- Fix arrow size on new `Carousel` buttons.
+
+## [2.63.0] - 2020-04-07
+
+Features:
+- Add new Dropdown in `dropdowns/v2`.
+- Add a11y options for new `Carousel` buttons.
+
+Fixes:
+- Revert old Dropdown.
+- Fix arrow size on new `Carousel` buttons.
 
 ## [2.62.0] - 2020-04-06